training/requirements.txt
# ML Training Requirements
# Install with: pip install -r requirements.txt

# Core ML
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0
opencv-python>=4.8.0

# YOLOv8
ultralytics>=8.0.0

# Model training
tensorflow>=2.13.0
keras>=2.13.0

# ArcFace (for feature extraction)
# pip install git+https://github.com/timesler/facenet-pytorch.git

# Conversion
tfjs-graph-converter>=0.9.0
onnxruntime>=1.15.0

# Data processing
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
albumentations>=1.3.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.23.0
